Переименован неймспейс. Инициализация ``` InputWrapper.cs ``` в PersistentObjectSpawner.cs.
