+++ Assets/Game/Core/Core.prefab
@@ -259,7 +259,6 @@ GameObject:
   m_Component:
   - component: {fileID: 8584269290858216682}
   - component: {fileID: 5529679516613883119}
-  - component: {fileID: 5275465452607436097}
   m_Layer: 0
   m_Name: Core
   m_TagString: Untagged
@@ -302,18 +301,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _persistentObjectsPrefab: {fileID: 1450072215902362737, guid: 82723e7b8a33c464893eddc285467fab,
     type: 3}
---- !u!114 &5275465452607436097
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5199896810855123400}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d2a1483ddb8dcce42b46a6bdbd88cdb1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &6741567287475702453
 GameObject:
   m_ObjectHideFlags: 0
