--- Assets/Scripts/Input/CustomInputSystemInitializer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d2a1483ddb8dcce42b46a6bdbd88cdb1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: -1
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
