+++ Assets/Scripts/InventorySystem/UI/InventoryItemIcon.cs
@@ -12,7 +12,7 @@ namespace FlavorfulStory.InventorySystem.UI
         /// <summary> Gameobject-контейнер количества предметов. </summary>
         [SerializeField] private GameObject _itemNumberContainer;
         
-        /// <summary> Текст количества предметов.</summary>
+        /// <summary> Текст количества предметов. </summary>
         [SerializeField] private TMP_Text _itemNumberText;
         
         /// <summary> Установить предмет инвентаря. </summary>
