+++ Assets/Scripts/UI/Tab.cs
@@ -1,5 +1,5 @@
 using System;
-using FlavorfulStory.Input;
+using FlavorfulStory.InputSystem;
 using UnityEngine;
 
 namespace FlavorfulStory.UI
@@ -10,26 +10,26 @@ namespace FlavorfulStory.UI
         /// <summary> Кнопка вкладки. </summary>
         [SerializeField] private TabButton _tabButton;
 
-        ///<summary> Контент, который отображается при активной вкладке. </summary>
+        /// <summary> Контент, который отображается при активной вкладке. </summary>
         [SerializeField] private GameObject _tabContent;
         
-        ///<summary> Индекс вкладки в наборе вкладок. </summary>
+        /// <summary> Индекс вкладки в наборе вкладок. </summary>
         private int _index;
 
-        ///<summary> Кнопка ввода, связанная с данной вкладкой. </summary>
+        /// <summary> Кнопка ввода, связанная с данной вкладкой. </summary>
         [field: SerializeField] public InputButton InputButton { get; private set; }
 
-        ///<summary> Событие, вызываемое при выборе вкладки. </summary>
+        /// <summary> Событие, вызываемое при выборе вкладки. </summary>
         ///<remarks> Передает индекс выбранной вкладки. </remarks>
         public event Action<int> OnTabSelected;
         
-        ///<summary> Инициализирует обработчик события клика по кнопке вкладки. </summary>
+        /// <summary> Инициализирует обработчик события клика по кнопке вкладки. </summary>
         private void Awake()
         {
             _tabButton.OnClick += SwitchTo;
         }
 
-        ///<summary> Переключает интерфейс на данную вкладку. </summary>
+        /// <summary> Переключает интерфейс на данную вкладку. </summary>
         ///<remarks> Вызывает событие выбора вкладки и активирует её контент. </remarks>
         private void SwitchTo()
         {
@@ -37,7 +37,7 @@ namespace FlavorfulStory.UI
             Select();
         }
 
-        ///<summary> Активирует вкладку и отображает её контент. </summary>
+        /// <summary> Активирует вкладку и отображает её контент. </summary>
         public void Select()
         {
             _tabButton.IsActive = true;
@@ -45,7 +45,7 @@ namespace FlavorfulStory.UI
             _tabContent.SetActive(true);
         }
 
-        ///<summary> Деактивирует вкладку и скрывает её контент. </summary>
+        /// <summary> Деактивирует вкладку и скрывает её контент. </summary>
         public void ResetSelection()
         {
             _tabButton.IsActive = false;
@@ -53,7 +53,7 @@ namespace FlavorfulStory.UI
             _tabContent.SetActive(false);
         }
 
-        ///<summary> Устанавливает индекс вкладки. </summary>
+        /// <summary> Устанавливает индекс вкладки. </summary>
         ///<param name="index"> Числовой индекс вкладки. </param>
         public void SetIndex(int index) => _index = index;
     }
